Create handling.meta merger for modding GTA V
This was created solely to make it easier to add vehicle mods to GTA V. This file compares lines of code against one another and outputs the appropriate line to mergedHandling.meta, as well as creating a backup of the handling.meta currently in use. 

In order to properly use this program, one will need the filepaths to: the original handling.meta (unmodded from Rockstar), the handling.meta currently in use in the game directory (although it will need to be extracted with OpenIV first to be able to be read/accessed), and the new handling.meta.

This is beta 0.1 i guess, as it is my first program written for something other than class and has only been tested by me. It has worked for me just fine, although much more testing is required. I am unsure whether or not I will update this, as I would like to see if anyone but myself will ever use it. Feel free to alter/improve as you see fit, just please give credit if using this as a base/template for alterations/improvements. 

I am not responsible for any loss of data resulting from the use of this program. Back up your data and use at your own risk.